refactor: rename `init env` cmd to `init` and remove login cmd
This change folds the login handler into the prepareEnvironment
handler. The getEnvironmentStatus handler can now indicate if
prepareEnvironment will need user input, in which case the framework
may error out and guide the user to run `garden init`. This allows
providers to transparently initialize in cases where they don't need
user input, but retain the flexibility to ask for user input when
needed (e.g. for logins, asking for sudo permissions etc.).
