+++ garden-cli/src/plugins/kubernetes/actions.ts
@@ -6,25 +6,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import * as Bluebird from "bluebird"
-import * as inquirer from "inquirer"
-import * as Joi from "joi"
 import * as split from "split"
-import { uniq } from "lodash"
 import moment = require("moment")
 
-import { DeploymentError, NotFoundError, TimeoutError, ConfigurationError } from "../../exceptions"
-import { GetServiceLogsResult, LoginStatus } from "../../types/plugin/outputs"
-import { RunResult, TestResult } from "../../types/plugin/outputs"
+import { DeploymentError, ConfigurationError } from "../../exceptions"
+import { GetServiceLogsResult, RunResult, TestResult } from "../../types/plugin/outputs"
 import {
-  PrepareEnvironmentParams,
-  CleanupEnvironmentParams,
   ExecInServiceParams,
-  GetEnvironmentStatusParams,
   GetServiceLogsParams,
   GetServiceOutputsParams,
   GetTestResultParams,
-  PluginActionParamsBase,
   RunModuleParams,
   TestModuleParams,
   DeleteServiceParams,
@@ -32,22 +23,15 @@ import {
 } from "../../types/plugin/params"
 import { ModuleVersion } from "../../vcs/base"
 import { ContainerModule, helpers, validateContainerModule } from "../container"
-import { deserializeValues, serializeValues, splitFirst, sleep } from "../../util/util"
-import { joiIdentifier } from "../../config/common"
+import { deserializeValues, serializeValues, splitFirst } from "../../util/util"
 import { KubeApi } from "./api"
-import {
-  getAppNamespace,
-  getMetadataNamespace,
-  getAllGardenNamespaces,
-} from "./namespace"
-import { KUBECTL_DEFAULT_TIMEOUT, kubectl } from "./kubectl"
+import { getAppNamespace, getMetadataNamespace } from "./namespace"
+import { kubectl } from "./kubectl"
 import { DEFAULT_TEST_TIMEOUT } from "../../constants"
-import { KubernetesProvider, name as providerName } from "./kubernetes"
+import { KubernetesProvider } from "./kubernetes"
 import { deleteContainerService, getContainerServiceStatus } from "./deployment"
 import { ServiceStatus } from "../../types/service"
 import { ValidateModuleParams } from "../../types/plugin/params"
-
-const MAX_STORED_USERNAMES = 5
 
 export async function validate(params: ValidateModuleParams<ContainerModule>) {
   const config = await validateContainerModule(params)
@@ -74,84 +58,6 @@ export async function validate(params: ValidateModuleParams<ContainerModule>) {
       ingressSpec.hostname = hostname
     }
   }
-}
-
-export async function getEnvironmentStatus({ ctx }: GetEnvironmentStatusParams) {
-  const context = ctx.provider.config.context
-
-  try {
-    // TODO: use API instead of kubectl (I just couldn't find which API call to make)
-    await kubectl(context).call(["version"])
-  } catch (err) {
-    // TODO: catch error properly
-    if (err.detail.output) {
-      throw new DeploymentError(
-        `Unable to connect to Kubernetes cluster. ` +
-        `Please make sure it is running, reachable and that you have the right context configured.`,
-        {
-          context,
-          kubectlOutput: err.detail.output,
-        },
-      )
-    }
-    throw err
-  }
-
-  await Bluebird.all([
-    getMetadataNamespace(ctx, ctx.provider),
-    getAppNamespace(ctx, ctx.provider),
-  ])
-
-  return {
-    ready: true,
-    detail: <any>{},
-  }
-}
-
-export async function prepareEnvironment({ }: PrepareEnvironmentParams) {
-  // this happens implicitly in the `getEnvironmentStatus()` function
-  return {}
-}
-
-export async function cleanupEnvironment({ ctx, logEntry }: CleanupEnvironmentParams) {
-  const api = new KubeApi(ctx.provider)
-  const namespace = await getAppNamespace(ctx, ctx.provider)
-  const entry = logEntry && logEntry.info({
-    section: "kubernetes",
-    msg: `Deleting namespace ${namespace} (this may take a while)`,
-    status: "active",
-  })
-
-  try {
-    // Note: Need to call the delete method with an empty object
-    // TODO: any cast is required until https://github.com/kubernetes-client/javascript/issues/52 is fixed
-    await api.core.deleteNamespace(namespace, <any>{})
-  } catch (err) {
-    entry && entry.setError(err.message)
-    const availableNamespaces = await getAllGardenNamespaces(api)
-    throw new NotFoundError(err, { namespace, availableNamespaces })
-  }
-
-  // Wait until namespace has been deleted
-  const startTime = new Date().getTime()
-  while (true) {
-    await sleep(2000)
-
-    const nsNames = await getAllGardenNamespaces(api)
-    if (!nsNames.includes(namespace)) {
-      break
-    }
-
-    const now = new Date().getTime()
-    if (now - startTime > KUBECTL_DEFAULT_TIMEOUT * 1000) {
-      throw new TimeoutError(
-        `Timed out waiting for namespace ${namespace} delete to complete`,
-        { namespace },
-      )
-    }
-  }
-
-  return {}
 }
 
 export async function deleteService(params: DeleteServiceParams): Promise<ServiceStatus> {
@@ -396,96 +302,6 @@ export async function getServiceLogs(
       resolve({})
     })
   })
-}
-
-export async function getLoginStatus({ ctx }: PluginActionParamsBase): Promise<LoginStatus> {
-  const localConfig = await ctx.localConfigStore.get()
-  let currentUsername
-  if (localConfig.kubernetes) {
-    currentUsername = localConfig.kubernetes.username
-  }
-  return { loggedIn: !!currentUsername }
-}
-
-export async function login({ ctx, logEntry }: PluginActionParamsBase): Promise<LoginStatus> {
-  const entry = logEntry && logEntry.info({ section: "kubernetes", msg: "Logging in..." })
-  const localConfig = await ctx.localConfigStore.get()
-
-  let currentUsername
-  let prevUsernames: Array<string> = []
-
-  if (localConfig.kubernetes) {
-    currentUsername = localConfig.kubernetes.username
-    prevUsernames = localConfig.kubernetes["previous-usernames"] || []
-  }
-
-  if (currentUsername) {
-    entry && entry.setDone({
-      symbol: "info",
-      msg: `Already logged in as user ${currentUsername}`,
-    })
-
-    return { loggedIn: true }
-  }
-
-  const promptName = "username"
-  const newUserOption = "Add new user"
-  type Ans = { [promptName]: string }
-  let ans: Ans
-
-  const inputPrompt = async () => {
-    return inquirer.prompt({
-      name: promptName,
-      message: "Enter username",
-      validate: input => {
-        try {
-          Joi.attempt(input.trim(), joiIdentifier())
-        } catch (err) {
-          return `Invalid username, please try again\nError: ${err.message}`
-        }
-        return true
-      },
-    })
-  }
-  const choicesPrompt = async () => {
-    return inquirer.prompt({
-      name: promptName,
-      type: "list",
-      message: "Log in as...",
-      choices: [...prevUsernames, new inquirer.Separator(), newUserOption],
-    })
-  }
-  if (prevUsernames.length > 0) {
-    ans = await choicesPrompt() as Ans
-    if (ans.username === newUserOption) {
-      ans = await inputPrompt() as Ans
-    }
-  } else {
-    ans = await inputPrompt() as Ans
-  }
-
-  const username = ans.username.trim()
-  const newPrevUsernames = uniq([...prevUsernames, username].slice(-MAX_STORED_USERNAMES))
-
-  await ctx.localConfigStore.set([
-    { keyPath: [providerName, "username"], value: username },
-    { keyPath: [providerName, "previous-usernames"], value: newPrevUsernames },
-  ])
-
-  return { loggedIn: true }
-}
-
-export async function logout({ ctx, logEntry }: PluginActionParamsBase): Promise<LoginStatus> {
-  const entry = logEntry && logEntry.info({ section: "kubernetes", msg: "Logging out..." })
-  const localConfig = await ctx.localConfigStore.get()
-  const k8sConfig = localConfig.kubernetes || {}
-  if (k8sConfig.username) {
-    await ctx.localConfigStore.delete([providerName, "username"])
-    entry && entry.setSuccess("Logged out")
-  } else {
-    entry && entry.setSuccess("Already logged out")
-  }
-  return { loggedIn: false }
 }
 
 function getTestResultKey(module: ContainerModule, testName: string, version: ModuleVersion) {
