+++ garden-cli/src/plugins/kubernetes/local.ts
@@ -8,13 +8,9 @@
 
 import * as execa from "execa"
 import { safeLoad } from "js-yaml"
-import { every, values } from "lodash"
 import * as Joi from "joi"
 import { join } from "path"
-import { PluginError } from "../../exceptions"
 import { GardenPlugin } from "../../types/plugin/plugin"
-import { GetEnvironmentStatusParams, PrepareEnvironmentParams } from "../../types/plugin/params"
-import { getEnvironmentStatus, prepareEnvironment } from "./actions"
 import { validate } from "../../config/common"
 import {
   gardenPlugin as k8sPlugin,
@@ -22,12 +18,9 @@ import {
   kubernetesConfigBase,
   KubernetesConfig,
 } from "./kubernetes"
-import { getSystemGarden, isSystemGarden } from "./system"
 import { readFile } from "fs-extra"
-import { LogEntry } from "../../logger/log-entry"
 import { homedir } from "os"
-import { helm } from "./helm"
-import { PluginContext } from "../../plugin-context"
+import { getLocalEnvironmentStatus, prepareLocalEnvironment } from "./init"
 
 // TODO: split this into separate plugins to handle Docker for Mac and Minikube
 
@@ -35,81 +28,6 @@ import { PluginContext } from "../../plugin-context"
 // and none is explicitly configured in the garden.yml
 const supportedContexts = ["docker-for-desktop", "minikube"]
 const kubeConfigPath = join(homedir(), ".kube", "config")
-
-// extend the environment configuration to also set up an ingress controller and dashboard
-export async function getLocalEnvironmentStatus(
-  { ctx, logEntry }: GetEnvironmentStatusParams,
-) {
-  const status = await getEnvironmentStatus({ ctx, logEntry })
-
-  if (!isSystemGarden(ctx.provider)) {
-    const sysGarden = await getSystemGarden(ctx.provider)
-    const sysStatus = await sysGarden.actions.getStatus()
-
-    status.detail.systemReady = sysStatus.providers[ctx.provider.config.name].ready &&
-      every(values(sysStatus.services).map(s => s.state === "ready"))
-    // status.detail.systemServicesStatus = sysStatus.services
-  }
-
-  status.ready = every(values(status.detail))
-
-  return status
-}
-
-async function configureSystemEnvironment(
-  { ctx, force, logEntry }:
-    { ctx: PluginContext, force: boolean, logEntry?: LogEntry },
-) {
-  const provider = ctx.provider
-  const sysGarden = await getSystemGarden(provider)
-  const sysCtx = sysGarden.getPluginContext(provider.name)
-
-  // TODO: need to add logic here to wait for tiller to be ready
-  await helm(sysCtx.provider,
-    "init", "--wait",
-    "--service-account", "default",
-    "--upgrade",
-  )
-
-  const sysStatus = await getEnvironmentStatus({
-    ctx: sysCtx,
-    logEntry,
-  })
-
-  await prepareEnvironment({
-    ctx: sysCtx,
-    force,
-    status: sysStatus,
-    logEntry,
-  })
-
-  // only deploy services if configured to do so (minikube bundles the required services as addons)
-  if (!provider.config._systemServices || provider.config._systemServices.length > 0) {
-    const results = await sysGarden.actions.deployServices({
-      serviceNames: provider.config._systemServices,
-    })
-
-    const failed = values(results.taskResults).filter(r => !!r.error).length
-
-    if (failed) {
-      throw new PluginError(`local-kubernetes: ${failed} errors occurred when configuring environment`, {
-        results,
-      })
-    }
-  }
-}
-
-async function configureLocalEnvironment(
-  { ctx, force, status, logEntry }: PrepareEnvironmentParams,
-) {
-  await prepareEnvironment({ ctx, force, status, logEntry })
-
-  if (!isSystemGarden(ctx.provider)) {
-    await configureSystemEnvironment({ ctx, force, logEntry })
-  }
-
-  return {}
-}
 
 async function getKubeConfig(): Promise<any> {
   try {
@@ -231,7 +149,7 @@ export async function gardenPlugin({ projectName, config, logEntry }): Promise<G
 
   // override the environment configuration steps
   plugin.actions!.getEnvironmentStatus = getLocalEnvironmentStatus
-  plugin.actions!.prepareEnvironment = configureLocalEnvironment
+  plugin.actions!.prepareEnvironment = prepareLocalEnvironment
 
   // no need to push before deploying locally
   delete plugin.moduleActions!.container.pushModule
